Move CWGraph
