+++ grakn-test/src/test/java/ai/grakn/test/graql/reasoner/AtomicTest.java
@@ -26,6 +26,7 @@ import ai.grakn.concept.TypeLabel;
 import ai.grakn.exception.GraqlQueryException;
 import ai.grakn.graphs.CWGraph;
 import ai.grakn.graql.MatchQuery;
+import ai.grakn.graql.Graql;
 import ai.grakn.graql.Var;
 import ai.grakn.graql.admin.Answer;
 import ai.grakn.graql.admin.Conjunction;
